Solve unknown bugs

+++ app/src/main/java/cn/eejing/ejcolorflower/model/request/GoodsListBean.java
@@ -12,7 +12,7 @@ public class GoodsListBean {
     /**
      * code : 1
      * message : 操作成功
-     * data : [{"goods_id":1,"name":"料包","money":50,"sold":"500","image":"http://60.205.226.109/uploads/goods/20180511/9f6556b744db2c89368838abf7588564.jpg","stock":1},{"goods_id":2,"name":"炫彩无敌小火花","money":500,"sold":"3000","image":"http://60.205.226.109/uploads/goods/20180511/4c224386671fa97601258ebb855e127d.jpg","stock":1},{"goods_id":3,"name":"花料","money":5000,"sold":"50000","image":"http://60.205.226.109/uploads/goods/20180511/7b1c1cb0b5573ea1b14354b232eae3bc.jpg","stock":0},{"goods_id":4,"name":"无敌小火花","money":50,"sold":"400","image":"http://60.205.226.109/uploads/goods/20180511/4528d302c61f5a05b78fda02ff3bc444.jpg","stock":1},{"goods_id":5,"name":"小旋风彩彩花","money":789,"sold":"987","image":"http://60.205.226.109/uploads/goods/20180511/58ccf875a1f9216bccc44fe728a9e515.jpg","stock":1},{"goods_id":7,"name":"小旋风彩彩花","money":520,"sold":"3000","image":"http://60.205.226.109/uploads/goods\\20180511\\67c23dff5ba9f38208e7fdc72b9d7306.jpg","stock":1}]
+     * data : [{"goods_id":1,"name":"炫彩烟花机","money":0.01,"sold":"0","image":"http://60.205.226.109/uploads/goods/20180521/2ee6635624f23464e6f61ca86172da96.jpg","stock":1},{"goods_id":2,"name":"炫彩烟花料(大)","money":65,"sold":"0","image":"http://60.205.226.109/uploads/goods/20180521/4341a22a4401a936794e71ea53be7cc4.jpg","stock":1},{"goods_id":3,"name":"炫彩烟花料(中)","money":35,"sold":"0","image":"http://60.205.226.109/uploads/goods/20180521/d98826a9cbde167b070a041c56ec6a02.jpg","stock":1},{"goods_id":4,"name":"炫彩烟花机","money":555,"sold":"0","image":"http://60.205.226.109/uploads/goods/20180521/0a7422d413c023381c01b2e9fb87d540.jpg","stock":1}]
      */
 
     private int code;
@@ -46,16 +46,16 @@ public class GoodsListBean {
     public static class DataBean {
         /**
          * goods_id : 1
-         * name : 料包
-         * money : 50
-         * sold : 500
-         * image : http://60.205.226.109/uploads/goods/20180511/9f6556b744db2c89368838abf7588564.jpg
+         * name : 炫彩烟花机
+         * money : 0.01
+         * sold : 0
+         * image : http://60.205.226.109/uploads/goods/20180521/2ee6635624f23464e6f61ca86172da96.jpg
          * stock : 1
          */
 
         private int goods_id;
         private String name;
-        private int money;
+        private double money;
         private String sold;
         private String image;
         private int stock;
@@ -76,11 +76,11 @@ public class GoodsListBean {
             this.name = name;
         }
 
-        public int getMoney() {
+        public double getMoney() {
             return money;
         }
 
-        public void setMoney(int money) {
+        public void setMoney(double money) {
             this.money = money;
         }
